Revised compartment to separate item pointer and check out pointer.
Made compartment a struct holding a pointer to an item and a pointer to another struct checkout info. So that if item ptr and check out can initially point to null and can create methods easier based around nullptr checks. shelf then is a class that contains 15 compartments and the overloaded[] operator allows you to directly obtain the item, because it returns a pointer reference. This allows for this call in main to be valid cout << libraryInventory[0][0]; without having to use any extra getter methods. The other major fix, prof mention checkout info not having to have methods to check for check out status because it could be done by checking for null which is why i made it a pointer in compartment to a check out info struct that contains the string member for name and string name for description.